server/requirements.txt
tiktoken
chainlit
pypdf2
langchain
python-dotenv
openai
faiss-cpu
huggingface_hub
InstructorEmbedding
sentence_transformers
tiktoken
chainlit
flask
pymongo
langdetect 
fpdf 
transformers 
huggingface 
pypdf2 
pdfminer.six 
PyMuPDF
reportlab